Split out cards, relics, and potions
