+++ events.py
@@ -4,12 +4,14 @@ import random
 from time import sleep
 from typing import Callable
 
+import card_catalog
+import displayer as view
+import generators as gen
 import items
+import potion_catalog
+import relic_catalog
 from ansi_tags import ansiprint
 from definitions import CardType, CombatTier, PlayerClass
-import effect_interface as ei
-import generators as gen
-import displayer as view
 from player import Player
 
 
@@ -146,7 +148,7 @@ def event_GoldenShrine(player):
             ansiprint("""Each time you strike the shrine, <yellow>gold</yellow> pours forth again and again!
 
 As you pocket the riches, something <red>weighs heavily on you.</red>""")
-            player.deck.append(items.Regret())
+            player.deck.append(card_catalog.Regret())
             ansiprint(f"{player.name} gained <curse>Regret</curse>")
             sleep(1.5)
             view.clear()
@@ -170,7 +172,7 @@ def event_Lab(player):
 Why do you know the name of all these tools? It doesn't matter, you take a look around.""")
     ansiprint("<bold>[Search]</bold> <green>Obtain 3 random potions</green>")
     input('Press enter > ')
-    gen.claim_potions(True, 3, player, items.create_all_potions())
+    gen.claim_potions(True, 3, player, potion_catalog.create_all_potions())
 
 # Won't add the Match and Keep event because i just don't know how to.
 
@@ -192,7 +194,7 @@ def event_OminousForge(player: Player):
             ansiprint('''You decide to see if you can find anything of use. After uncovering tarps, looking through boxes, and checking nooks and crannies, you find a dust covered <yellow>relic!</yellow>.
 
 Taking the relic, you can't shake a sudden feeling of <red>sharp pain</red> as you exit the hut. Maybe you disturbed some sort of spirit?''')
-            gen.claim_relics(False, player, 1, items.create_all_relics(), [items.WarpedTongs()])
+            gen.claim_relics(False, player, 1, relic_catalog.create_all_relics(), [items.WarpedTongs()])
             gen.card_rewards(CombatTier.NORMAL, False, player, card_catalog.create_all_cards(), [items.Pain()])
             break
     input('Press enter to continue > ')
@@ -290,7 +292,7 @@ You eye him suspiciously and consider your options...""")
         ansiprint("<bold>[Give Potion]</bold> <red>Lose a potion.</red> <green>Recieve a relic.</green> \n<bold>[Give Gold]</bold> <red>Lose a varying amount of gold.</red> <green>Recieve a relic.</green> \n<bold>[Give Card]</bold> <red>Lose a card.</red> <green>Recieve a relic.</green> \n<bold>[Attack]</bold> Nothing happens.")
         option = input('> ').lower()
         if 'give' in option:
-            relic_rewards = [relic for relic in items.create_all_relics() if relic.get('Rarity') in ('Common', 'Uncommon', 'Rare')]
+            relic_rewards = [relic for relic in relic_catalog.create_all_relics() if relic.get('Rarity') in ('Common', 'Uncommon', 'Rare')]
             if 'potion' in option:
                 if len(player.potions) == 0:
                     print("You don't have any potions.")
@@ -345,7 +347,7 @@ He runs away.''')
 
 
 def event_TheWomanInBlue(player):
-    valid_potions = {potion: stats for potion, stats in items.create_all_potions().items() if stats.get('Class') == player.player_class}
+    valid_potions = {potion: stats for potion, stats in potion_catalog.create_all_potions().items() if stats.get('Class') == player.player_class}
     while True:
         ansiprint('<bold>The Woman in Blue</bold>')
         sleep(0.8)
@@ -378,8 +380,8 @@ You exit the shop cautiously.''')
 
 
 def event_FaceTrader(player):
-    face_relics = [items.create_all_relics()['Cultist Headpiece'], items.create_all_relics()['Face of Cleric'],
-                   items.create_all_relics()['Ssserpent Head'], items.create_all_relics()['Gremlin Visage'], items.create_all_relics()["N'loth's Hungry Face"]]
+    face_relics = [relic_catalog.create_all_relics()['Cultist Headpiece'], relic_catalog.create_all_relics()['Face of Cleric'],
+                   relic_catalog.create_all_relics()['Ssserpent Head'], relic_catalog.create_all_relics()['Gremlin Visage'], relic_catalog.create_all_relics()["N'loth's Hungry Face"]]
     while True:
         ansiprint('<bold>Face Trader</bold>\n')
         sleep(0.8)
@@ -410,7 +412,7 @@ During this, his mask falls off and shatters. Screaming, he quickly covers his f
 His face was completely blank.''')
             break
         if option == 'trade':
-            gen.claim_relics(False, player, 1, items.create_all_relics(), [random.choice(face_relics)], False)
+            gen.claim_relics(False, player, 1, relic_catalog.create_all_relics(), [random.choice(face_relics)], False)
             sleep(0.8)
             ansiprint('''<bold>Eerie Man</bold>: "For me? <italic>FOR ME?</italic> Oh yes.. Yes. Yes.. mmm..."
 
@@ -454,9 +456,9 @@ What do you do?''')
         if option == 'box':
             ansiprint('You grab the box. Inside you find a <yellow>relic</yellow>! \nHowever, you really craved the donut... \nYou are filled with sadness, but mostly <red>regret</red>.')
             sleep(1.3)
-            gen.claim_relics(False, player, 1, items.create_all_relics(), None, False)
-            player.deck.append(items.Regret())
-            ansiprint(f'You obtained <magenta>Regret</magenta> | {items.Regret().info}')
+            gen.claim_relics(False, player, 1, relic_catalog.create_all_relics(), None, False)
+            player.deck.append(card_catalog.Regret())
+            ansiprint(f'You obtained <magenta>Regret</magenta> | {card_catalog.Regret().info}')
         sleep(1.5)
         view.clear()
     input('Press enter to leave > ')
@@ -525,7 +527,7 @@ You're sure you don't see any traps nearby.\n""")
             view.clear()
             continue
         if option == 'take':
-            gen.claim_relics(False, player, 1, items.create_all_relics(), [items.GoldenIdol], False)
+            gen.claim_relics(False, player, 1, relic_catalog.create_all_relics(), [items.GoldenIdol], False)
             ansiprint("""As you grab the idol and stow it away, a giant boulder smashes through the ceiling into the ground next to you.
 
 You realize that the floor is slanted downwards as the boulder starts to roll towards you.""")
